+++ src/Aiursoft.StatHub.SDK/Models/DiskSpaceInfo.cs
@@ -3,6 +3,6 @@ namespace Aiursoft.StatHub.SDK.Models;
 public class DiskSpaceInfo
 {
     public string Name { get; set; } = null!;
-    public int Total { get; set; }
-    public int Used { get; set; }
+    public double Total { get; set; }
+    public double Used { get; set; }
 }
